# perf(nhcb): reuse calculated metric name and hash