The benchmark showed little improvement. I'll run prombench to see if it improved over https://github.com/prometheus/prometheus/pull/15467#issuecomment-2569844862
```
                                                         │   main.txt   │               pr.txt               │
                                                         │    sec/op    │    sec/op     vs base              │
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2        62.53µ ± 10%   62.14µ ± 13%       ~ (p=0.589 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2   96.58µ ± 19%   86.50µ ± 17%       ~ (p=0.310 n=6)
geomean                                                    77.71µ         73.31µ        -5.66%

                                                         │   main.txt    │               pr.txt                │
                                                         │      B/s      │      B/s       vs base              │
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2        68.29Mi ± 11%   68.71Mi ± 14%       ~ (p=0.589 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2   41.47Mi ± 24%   46.42Mi ± 15%       ~ (p=0.310 n=6)
geomean                                                    53.22Mi         56.48Mi        +6.13%

                                                         │   main.txt   │               pr.txt               │
                                                         │     B/op     │     B/op      vs base              │
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2        14.05Ki ± 0%   14.11Ki ± 0%  +0.44% (p=0.002 n=6)
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2   22.49Ki ± 0%   22.31Ki ± 0%  -0.76% (p=0.002 n=6)
geomean                                                    17.77Ki        17.74Ki       -0.16%

                                                         │  main.txt  │               pr.txt               │
                                                         │ allocs/op  │ allocs/op   vs base                │
Parse/data=omtestdata.txt/parser=omtext_with_nhcb-2        237.0 ± 0%   237.0 ± 0%       ~ (p=1.000 n=6) ¹
Parse/data=omhistogramdata.txt/parser=omtext_with_nhcb-2   355.0 ± 0%   347.0 ± 0%  -2.25% (p=0.002 n=6)
geomean                                                    290.1        286.8       -1.13%
¹ all samples are equal

```