@bwplotka 2025-01-27
As per https://github.com/prometheus/test-infra/tree/bench/nhcbconvert/prombench/manifests/prombench we might another custom benchmark that will make BOTH A & B Prometheus to use NHCB to compare improvements.

Hard to tell between this PR prombench:

<img width="595" alt="image" src="https://github.com/user-attachments/assets/c3b6bf3d-3777-4581-910e-5c7bb244fdca" />

<img width="335" alt="image" src="https://github.com/user-attachments/assets/4f1ba534-ebd6-4d65-965e-596687c58aa2" />

..and https://github.com/prometheus/prometheus/pull/15467#issuecomment-2569844862