@bwplotka 2025-01-27
Isn't this PR actually making things worse? Even macro benchmark is stating it allocs more per second which we want to optimize for. 

I think we need to combine converter to Text parsers directly, that feels the most impactful to memory IMO at this point. I can create custom benchmark for us if you want.