@bwplotka 2025-01-27
/prombench main --bench.version=bench/nhcb