@bwplotka 2025-02-05
For the record on the prombench run, I don't see a difference in prombench with nhcb:

![image](https://github.com/user-attachments/assets/1e20d72a-a2bc-4931-bde8-dee398f429a9)

![image](https://github.com/user-attachments/assets/d1684e6b-a501-413f-8a3e-d2b2fa6cfc32)

![image](https://github.com/user-attachments/assets/7816ec74-860e-4562-876a-2ddb8f0ca0e8)