@bwplotka 2025-02-05
/prombench main

Testing prombench from the PR rebased on older commits that worked before (https://github.com/prometheus/test-infra/issues/832)